+++ Package.swift
@@ -4,6 +4,7 @@ import PackageDescription
 
 let package = Package(
     name: "KeyboardKit",
+    defaultLocalization: "en",
     platforms: [
         .iOS(.v13)
     ],